refactor: NFTView and SubdomainView to pages
