+++ src/features/subdomain-view/SubdomainTableCard/SubdomainTableCard.tsx
@@ -6,15 +6,15 @@ import { ethers } from 'ethers';
 import { TokenPriceInfo } from '@zero-tech/zns-sdk';
 
 //- Hook Imports
-import { useDomainMetadata } from '../../../../lib/hooks/useDomainMetadata';
-import { useDomainMetrics } from '../../../../lib/hooks/useDomainMetrics';
-import { formatEthers, formatNumber } from '../../../../lib/util/number/number';
-import { useBuyNowPrice } from '../../../../lib/hooks/useBuyNowPrice';
+import { useDomainMetadata } from '../../../lib/hooks/useDomainMetadata';
+import { useDomainMetrics } from '../../../lib/hooks/useDomainMetrics';
+import { formatEthers, formatNumber } from '../../../lib/util/number/number';
+import { useBuyNowPrice } from '../../../lib/hooks/useBuyNowPrice';
 
 //- Component Imports
-import { PlaceBidButton } from '../../../../features/place-bid';
-import { BuyNowButton } from '../../../../features/buy-now';
-import TableCard from '../../../ui/TableCard/TableCard';
+import { PlaceBidButton } from '../../place-bid';
+import { BuyNowButton } from '../../buy-now';
+import TableCard from '../../ui/TableCard/TableCard';
 
 //-Styles Imports
 import styles from './SubdomainTableCard.module.scss';
