+++ src/features/subdomain-view/SubdomainTableRow/SubdomainTableRow.tsx
@@ -6,14 +6,14 @@ import { ethers } from 'ethers';
 import { TokenPriceInfo } from '@zero-tech/zns-sdk';
 
 //- Hook Imports
-import { useDomainMetadata } from '../../../../lib/hooks/useDomainMetadata';
-import { useDomainMetrics } from '../../../../lib/hooks/useDomainMetrics';
-import { formatEthers, formatNumber } from '../../../../lib/util/number/number';
-import { useBuyNowPrice } from '../../../../lib/hooks/useBuyNowPrice';
+import { useDomainMetadata } from '../../../lib/hooks/useDomainMetadata';
+import { useDomainMetrics } from '../../../lib/hooks/useDomainMetrics';
+import { formatEthers, formatNumber } from '../../../lib/util/number/number';
+import { useBuyNowPrice } from '../../../lib/hooks/useBuyNowPrice';
 
 //- Component Imports
-import { PlaceBidButton } from '../../../../features/place-bid';
-import { BuyNowButton } from '../../../../features/buy-now';
+import { PlaceBidButton } from '../../place-bid';
+import { BuyNowButton } from '../../buy-now';
 
 type SubdomainTableRowProps = {
 	domainId: string;
